Refactor Role Based Scope validation
