+++ components/wso2is.key.manager.core/src/main/java/org/wso2/is/key/manager/core/tokenmgt/issuers/RoleBasedScopesIssuer.java
@@ -18,6 +18,8 @@
  */
 package org.wso2.is.key.manager.core.tokenmgt.issuers;
 
+import com.nimbusds.jwt.JWTClaimsSet;
+import com.nimbusds.jwt.SignedJWT;
 import org.apache.axis2.util.JavaUtils;
 import org.apache.commons.lang.ArrayUtils;
 import org.apache.commons.lang.StringUtils;
@@ -29,6 +31,12 @@ import org.osgi.service.component.annotations.Component;
 import org.wso2.carbon.identity.application.authentication.framework.model.AuthenticatedUser;
 import org.wso2.carbon.identity.application.authentication.framework.util.FrameworkUtils;
 import org.wso2.carbon.identity.application.common.model.ClaimMapping;
+import org.wso2.carbon.identity.application.common.model.FederatedAuthenticatorConfig;
+import org.wso2.carbon.identity.application.common.model.IdentityProvider;
+import org.wso2.carbon.identity.application.common.model.PermissionsAndRoleConfig;
+import org.wso2.carbon.identity.application.common.model.RoleMapping;
+import org.wso2.carbon.identity.application.common.util.IdentityApplicationConstants;
+import org.wso2.carbon.identity.application.common.util.IdentityApplicationManagementUtil;
 import org.wso2.carbon.identity.base.IdentityConstants;
 import org.wso2.carbon.identity.core.util.IdentityTenantUtil;
 import org.wso2.carbon.identity.core.util.IdentityUtil;
@@ -48,18 +56,23 @@ import org.wso2.carbon.identity.oauth2.dao.OAuthTokenPersistenceFactory;
 import org.wso2.carbon.identity.oauth2.dto.OAuth2TokenValidationRequestDTO;
 import org.wso2.carbon.identity.oauth2.grant.jwt.JWTConstants;
 import org.wso2.carbon.identity.oauth2.model.AccessTokenDO;
+import org.wso2.carbon.identity.oauth2.model.RequestParameter;
 import org.wso2.carbon.identity.oauth2.model.ResourceScopeCacheEntry;
 import org.wso2.carbon.identity.oauth2.token.OAuthTokenReqMessageContext;
 import org.wso2.carbon.identity.oauth2.validators.OAuth2TokenValidationMessageContext;
 import org.wso2.carbon.identity.oauth2.validators.scope.ScopeValidator;
+import org.wso2.carbon.idp.mgt.IdentityProviderManagementException;
+import org.wso2.carbon.idp.mgt.IdentityProviderManager;
 import org.wso2.carbon.user.api.UserStoreException;
 import org.wso2.carbon.user.api.UserStoreManager;
+import org.wso2.carbon.user.core.UserCoreConstants;
 import org.wso2.carbon.user.core.service.RealmService;
 import org.wso2.carbon.user.core.util.UserCoreUtil;
 import org.wso2.carbon.utils.multitenancy.MultitenantUtils;
 import org.wso2.is.key.manager.core.tokenmgt.handlers.ResourceConstants;
 import org.wso2.is.key.manager.core.tokenmgt.util.TokenMgtUtil;
 
+import java.text.ParseException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Iterator;
@@ -92,8 +105,7 @@ public class RoleBasedScopesIssuer extends AbstractScopesIssuer implements Scope
     private static final String SCOPE_VALIDATOR_NAME = "Role based scope validator";
     private static final String OPENID = "openid";
     protected boolean rejectUnregisteredScopes = false;
-    List<String> allowed;
-
+    private IdentityProvider identityProvider = null;
     // set role based scopes issuer as the default
     private static final String ISSUER_PREFIX = "default";
     OAuthServerConfiguration oAuthServerConfiguration = OAuthServerConfiguration.getInstance();
@@ -201,9 +213,6 @@ public class RoleBasedScopesIssuer extends AbstractScopesIssuer implements Scope
         }
 
         try {
-            /*User authzUser = accessTokenDO.getAuthzUser();
-            int tenantId = getTenantId(authzUser);
-            String[] userRoles = getUserRoles(authzUser);*/
             AuthenticatedUser authenticatedUser = OAuthUtil
                     .getAuthenticatedUser(oAuth2TokenValidationMessageContext.getResponseDTO().getAuthorizedUser());
             String clientId = accessTokenDO.getConsumerKey();
@@ -361,12 +370,14 @@ public class RoleBasedScopesIssuer extends AbstractScopesIssuer implements Scope
             String isRetrieveRolesFromUserStoreForScopeValidation = System
                     .getProperty(ResourceConstants.RETRIEVE_ROLES_FROM_USERSTORE_FOR_SCOPE_VALIDATION);
             if (GrantType.SAML20_BEARER.toString().equals(grantType) && Boolean.parseBoolean(isSAML2Enabled)) {
+                authenticatedUser.setUserStoreDomain("FEDERATED");
+                tokReqMsgCtx.setAuthorizedUser(authenticatedUser);
                 Assertion assertion = (Assertion) tokReqMsgCtx.getProperty(ResourceConstants.SAML2_ASSERTION);
                 userRoles = getRolesFromAssertion(assertion);
             } else if (JWTConstants.OAUTH_JWT_BEARER_GRANT_TYPE.equals(grantType) && !(Boolean
                     .parseBoolean(isRetrieveRolesFromUserStoreForScopeValidation))) {
-                AuthenticatedUser user = tokReqMsgCtx.getAuthorizedUser();
-                Map<ClaimMapping, String> userAttributes = user.getUserAttributes();
+                configureForJWTGrant(tokReqMsgCtx);
+                Map<ClaimMapping, String> userAttributes = authenticatedUser.getUserAttributes();
                 if (tokReqMsgCtx.getProperty(ResourceConstants.ROLE_CLAIM) != null) {
                     userRoles = getRolesFromUserAttribute(userAttributes,
                             tokReqMsgCtx.getProperty(ResourceConstants.ROLE_CLAIM).toString());
@@ -551,5 +562,193 @@ public class RoleBasedScopesIssuer extends AbstractScopesIssuer implements Scope
      */
     protected String[] getRolesFromAssertion(Assertion assertion) {
         return TokenMgtUtil.getRolesFromAssertion(assertion);
+    }
+
+    protected void configureForJWTGrant(OAuthTokenReqMessageContext tokReqMsgCtx) {
+
+        SignedJWT signedJWT = null;
+        JWTClaimsSet claimsSet = null;
+        String[] roles = null;
+        try {
+            signedJWT = getSignedJWT(tokReqMsgCtx);
+        } catch (IdentityOAuth2Exception e) {
+            log.error("Couldn't retrieve signed JWT", e);
+        }
+        if (signedJWT != null) {
+            claimsSet = getClaimSet(signedJWT);
+        }
+        String jwtIssuer = claimsSet != null ? claimsSet.getIssuer() : null;
+        String tenantDomain = tokReqMsgCtx.getOauth2AccessTokenReqDTO().getTenantDomain();
+
+        try {
+            identityProvider = IdentityProviderManager.getInstance().getIdPByName(jwtIssuer, tenantDomain);
+            if (identityProvider != null) {
+                if (StringUtils.equalsIgnoreCase(identityProvider.getIdentityProviderName(), "default")) {
+                    identityProvider = this.getResidentIDPForIssuer(tenantDomain, jwtIssuer);
+                    if (identityProvider == null) {
+                        log.error("No Registered IDP found for the JWT with issuer name : " + jwtIssuer);
+                    }
+                }
+            } else {
+                log.error("No Registered IDP found for the JWT with issuer name : " + jwtIssuer);
+            }
+        } catch (IdentityProviderManagementException | IdentityOAuth2Exception e) {
+            log.error("Couldn't initiate identity provider instance", e);
+        }
+
+        try {
+            roles = claimsSet != null ?
+                    claimsSet.getStringArrayClaim(identityProvider.getClaimConfig().getRoleClaimURI()) :
+                    null;
+        } catch (ParseException e) {
+            log.error("Couldn't retrieve roles:", e);
+        }
+
+        List<String> updatedRoles = new ArrayList<>();
+        if (roles != null) {
+            for (String role : roles) {
+                String updatedRoleClaimValue = getUpdatedRoleClaimValue(identityProvider, role);
+                if (updatedRoleClaimValue != null) {
+                    updatedRoles.add(updatedRoleClaimValue);
+                } else {
+                    updatedRoles.add(role);
+                }
+            }
+        }
+        AuthenticatedUser user = tokReqMsgCtx.getAuthorizedUser();
+        Map<ClaimMapping, String> userAttributes = user.getUserAttributes();
+        String roleClaim = identityProvider.getClaimConfig().getRoleClaimURI();
+        if (roleClaim != null) {
+            userAttributes
+                    .put(ClaimMapping.build(roleClaim, roleClaim, null, false),
+                            updatedRoles.toString().replace(" ", ""));
+            tokReqMsgCtx.addProperty(ResourceConstants.ROLE_CLAIM, roleClaim);
+        }
+        user.setUserAttributes(userAttributes);
+        tokReqMsgCtx.setAuthorizedUser(user);
+    }
+
+    /**
+     * Method to parse the assertion and retrieve the signed JWT
+     *
+     * @param tokReqMsgCtx request
+     * @return SignedJWT object
+     * @throws IdentityOAuth2Exception exception thrown due to a parsing error
+     */
+    private SignedJWT getSignedJWT(OAuthTokenReqMessageContext tokReqMsgCtx) throws IdentityOAuth2Exception {
+        RequestParameter[] params = tokReqMsgCtx.getOauth2AccessTokenReqDTO().getRequestParameters();
+        String assertion = null;
+        SignedJWT signedJWT;
+        for (RequestParameter param : params) {
+            if (param.getKey().equals(JWTConstants.OAUTH_JWT_ASSERTION)) {
+                assertion = param.getValue()[0];
+                break;
+            }
+        }
+        if (StringUtils.isEmpty(assertion)) {
+            String errorMessage = "Error while retrieving assertion";
+            throw new IdentityOAuth2Exception(errorMessage);
+        }
+
+        try {
+            signedJWT = SignedJWT.parse(assertion);
+            if (log.isDebugEnabled()) {
+                log.debug(signedJWT);
+            }
+        } catch (ParseException e) {
+            String errorMessage = "Error while parsing the JWT.";
+            throw new IdentityOAuth2Exception(errorMessage, e);
+        }
+        return signedJWT;
+    }
+
+    /**
+     * Method to retrieve claims from the JWT
+     *
+     * @param signedJWT JWT token
+     * @return JWTClaimsSet Object
+     */
+    private JWTClaimsSet getClaimSet(SignedJWT signedJWT) {
+        JWTClaimsSet claimsSet = null;
+        try {
+            claimsSet = signedJWT.getJWTClaimsSet();
+        } catch (ParseException e) {
+            log.error("Error when trying to retrieve claimsSet from the JWT:", e);
+        }
+        return claimsSet;
+    }
+
+    private IdentityProvider getResidentIDPForIssuer(String tenantDomain, String jwtIssuer)
+            throws IdentityOAuth2Exception {
+        String issuer = "";
+
+        IdentityProvider residentIdentityProvider;
+        try {
+            residentIdentityProvider = IdentityProviderManager.getInstance().getResidentIdP(tenantDomain);
+        } catch (IdentityProviderManagementException var7) {
+            String errorMsg = String
+                    .format("Error while getting Resident Identity Provider of '%s' tenant.", tenantDomain);
+            throw new IdentityOAuth2Exception(errorMsg, var7);
+        }
+
+        FederatedAuthenticatorConfig[] fedAuthnConfigs = residentIdentityProvider.getFederatedAuthenticatorConfigs();
+        FederatedAuthenticatorConfig oauthAuthenticatorConfig = IdentityApplicationManagementUtil.
+                getFederatedAuthenticator(fedAuthnConfigs, "openidconnect");
+        if (oauthAuthenticatorConfig != null) {
+            issuer = IdentityApplicationManagementUtil.getProperty(oauthAuthenticatorConfig.
+                    getProperties(), "IdPEntityId").getValue();
+        }
+
+        return jwtIssuer.equals(issuer) ? residentIdentityProvider : null;
+    }
+
+    /**
+     * Check the retireved roles against the role mappings in the IDP and return the updated roles
+     *
+     * @param identityProvider      used to retrieve the role mappings
+     * @param currentRoleClaimValue current roles received through the token
+     * @return updated roles
+     */
+    private String getUpdatedRoleClaimValue(IdentityProvider identityProvider, String currentRoleClaimValue) {
+
+        if (StringUtils.equalsIgnoreCase(IdentityApplicationConstants.RESIDENT_IDP_RESERVED_NAME,
+                identityProvider.getIdentityProviderName())) {
+            return currentRoleClaimValue;
+        }
+        currentRoleClaimValue = currentRoleClaimValue.replace("\\/", "/")
+                .replace("[", "").replace("]", "")
+                .replace("\"", "");
+
+        PermissionsAndRoleConfig permissionAndRoleConfig = identityProvider.getPermissionAndRoleConfig();
+        if (permissionAndRoleConfig != null && org.apache.commons.lang3.ArrayUtils.isNotEmpty(
+                permissionAndRoleConfig.getRoleMappings())) {
+            String[] receivedRoles = currentRoleClaimValue.split(FrameworkUtils.getMultiAttributeSeparator());
+            List<String> updatedRoleClaimValues = new ArrayList<>();
+            String updatedLocalRole;
+            loop:
+            for (String receivedRole : receivedRoles) {
+                for (RoleMapping roleMapping : permissionAndRoleConfig.getRoleMappings()) {
+                    if (roleMapping.getRemoteRole().equals(receivedRole)) {
+                        updatedLocalRole = StringUtils.isEmpty(roleMapping.getLocalRole().getUserStoreId())
+                                ? roleMapping.getLocalRole().getLocalRoleName()
+                                : roleMapping.getLocalRole().getUserStoreId() + UserCoreConstants.DOMAIN_SEPARATOR
+                                + roleMapping.getLocalRole().getLocalRoleName();
+                        updatedRoleClaimValues.add(updatedLocalRole);
+                        continue loop;
+                    }
+                }
+                if (!OAuthServerConfiguration.getInstance().isReturnOnlyMappedLocalRoles()) {
+                    updatedRoleClaimValues.add(receivedRole);
+                }
+            }
+            if (!updatedRoleClaimValues.isEmpty()) {
+                return StringUtils.join(updatedRoleClaimValues, FrameworkUtils.getMultiAttributeSeparator());
+            }
+            return null;
+        }
+        if (!OAuthServerConfiguration.getInstance().isReturnOnlyMappedLocalRoles()) {
+            return currentRoleClaimValue;
+        }
+        return null;
     }
 }
